+++ src/main/java/com/bookStors/WookieBooks/repository/BooksRepository.java
@@ -4,5 +4,5 @@ import com.bookStors.WookieBooks.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BooksRepository extends JpaRepository<Book,Integer> {
+public interface BooksRepository extends JpaRepository<Book,Long> {
 }
